pyproject.toml
[project]
name = "inkypi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "numpy>=2.2.3",
    "openai>=1.65.4",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
    "pytz>=2025.1",
    "requests>=2.32.3",
    "urllib3>=2.3.0",
    "werkzeug>=3.1.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-env>=1.1.5",
    "ruff>=0.9.10",
]

[tool.pytest_env]
SRC_DIR = "../src"